cronjob_snapshot/src/requirements.txt
pandas
psycopg2-binary
black
pre-commit
SQLAlchemy
